[bluebinder] gbinder reader expects gsize for size and count. JB#42254

+++ bluebinder.c
@@ -420,7 +420,7 @@ bluebinder_callbacks_transact(
             }
 
         } else if (code == 2 || code == 3 || code == 4) {
-            unsigned int count, elemsize;
+            gsize count, elemsize;
             GBinderReader reader;
             const uint8_t *vec;
             uint8_t *packet;
